+++ offchain/dispatcher/src/setup.rs
@@ -5,7 +5,7 @@ use eth_state_client_lib::{
     config::SCConfig, error::StateServerError, BlockServer,
     GrpcStateFoldClient, StateServer,
 };
-use eth_state_fold_types::{ethereum_types::H256, BlockStreamItem};
+use eth_state_fold_types::{ethereum_types::U64, BlockStreamItem};
 use rollups_events::DAppMetadata;
 use snafu::{ensure, ResultExt};
 use tokio_stream::{Stream, StreamExt};
@@ -80,15 +80,10 @@ pub async fn create_context(
     dapp_metadata: DAppMetadata,
     metrics: DispatcherMetrics,
 ) -> Result<Context, DispatcherError> {
-    let dapp_deploy_block_hash = H256(
-        config
-            .blockchain_config
-            .dapp_deploy_block_hash
-            .clone()
-            .into_inner(),
-    );
+    let dapp_deployment_block_number =
+        U64::from(config.blockchain_config.dapp_deployment_block_number);
     let genesis_timestamp: u64 = block_server
-        .query_block(dapp_deploy_block_hash)
+        .query_block(dapp_deployment_block_number)
         .await
         .context(StateServerSnafu)?
         .timestamp